refactor structure to be internal util
